Refactor- [:feature:recent-transaction] Apply & Fix Detekt, Ktlint Rules
Jira Tasks - [MM-74](https://mifosforge.jira.com/browse/MM-74)
