+++ feature/recent-transaction/src/main/res/values/strings.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+    Copyright 2024 Mifos Initiative
+
+    This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0.
+    If a copy of the MPL was not distributed with this file,
+    You can obtain one at https://mozilla.org/MPL/2.0/.
+
+    See https://github.com/openMF/mobile-mobile/blob/master/LICENSE.md
+-->
+<resources>
+
+    <string name="recent_transactions">Recent Transactions</string>
+    <string name="no_transaction">No Transaction</string>
+    <string name="atm_icon">ATM Icon</string>
+    <string name="string_and_string">%1$s %2$s</string>
+</resources>
