--- feature/recent_transaction/build.gradle.kts
@@ -1,19 +1,0 @@
-plugins {
-    alias(libs.plugins.mifos.android.feature)
-    alias(libs.plugins.mifos.android.library.compose)
-}
-
-android {
-    namespace = "org.mifos.mobile.feature.recent_transaction"
-}
-
-dependencies {
-    implementation(projects.core.ui)
-    implementation(projects.core.common)
-    implementation(projects.core.model)
-    implementation(projects.core.data)
-
-    testImplementation(libs.junit)
-    androidTestImplementation(libs.androidx.test.ext.junit)
-    androidTestImplementation(libs.espresso.core)
-}
