+++ settings.gradle.kts
@@ -40,7 +40,7 @@ include(":feature:savings")
 include(":feature:qr")
 include(":feature:transfer-process")
 include(":feature:account")
-include(":feature:recent_transaction")
+include(":feature:recent-transaction")
 include(":feature:client-charge")
 include(":feature:third-party-transfer")
 include(":feature:help")
